Seokhyeon : Create UBattleWIdget / Edit AMyGameModeBase - Add CalculateTurn Function

+++ Source/VR12ForTheKing/Character/MyCharacter.h
@@ -51,27 +51,27 @@ private:
 
 	// Status
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Armor;
+	int32 Armor;
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Awareness;
+	int32 Awareness;
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Evasion;
+	int32 Evasion;
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Focus;
+	int32 Focus;
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Intelligence;
+	int32 Intelligence;
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Luck;
+	int32 Luck;
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Resistance;
+	int32 Resistance;
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Speed;
+	int32 Speed;
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Strength;
+	int32 Strength;
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Talent;
+	int32 Talent;
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
-	float Vitality;
+	int32 Vitality;
 
 	// Move Var
 	UPROPERTY(BlueprintReadWrite, Category = "Status", EditAnywhere, meta = (AllowPrivateAccess = true))
